Moving list_submissions to fileHandler. Updating logic for more efficient handling of filtering submissions based on publishable status. Moved unit tests to test_file_handler.
